cloud_functions/sustainable_sourcing_layers/requirements.txt
earthengine-api
flask
functions-framework
google-api-core
google-cloud-logging
requests
